Second commit - now able to return measurement from a device as json; Also implemented some interfaces to make the app more flexible

+++ src/main/java/com/magnanime/RestHomeAutomationRPiServer/DataModel/Device.java
@@ -16,5 +16,5 @@ public class Device implements Serializable {
     @Column(name = "name")
     private String name;
     @Column(name = "devicetype")
-    private long type;
+    private Integer type;
 }
